+++ packages/orq-rc/RELEASES.md
@@ -4939,3 +4939,13 @@ Based on:
 - [python v4.2.0-rc.56] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.56] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.56 - packages/orq-rc
+
+## 2026-01-22 05:48:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.693.0 (2.798.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.57] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.57] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.57 - packages/orq-rc